requirements.txt
django
MySQL-python
django-wysiwyg
django-ckeditor
django-registration-redux
django-tinymce
python-dotenv
bunch
django-dynamic-formset
